Note these results are on my laptop, so take with a grain of salt.
```
BenchmarkKVInsert1_Native-4    	    2000	    839340 ns/op
BenchmarkKVInsert1_SQL-4       	    2000	    649759 ns/op
BenchmarkKVInsert10_Native-4   	    2000	   1099658 ns/op
BenchmarkKVInsert10_SQL-4      	    1000	   1302572 ns/op
BenchmarkKVInsert100_Native-4  	     500	   3374758 ns/op
BenchmarkKVInsert100_SQL-4     	     200	   6999272 ns/op
BenchmarkKVInsert1000_Native-4 	      50	  25849605 ns/op
BenchmarkKVInsert1000_SQL-4    	      20	  72150309 ns/op
BenchmarkKVInsert10000_Native-4	       5	 299836340 ns/op
BenchmarkKVInsert10000_SQL-4   	       1	1045935749 ns/op
BenchmarkKVUpdate1_Native-4    	    2000	   1143479 ns/op
BenchmarkKVUpdate1_SQL-4       	    2000	   1006438 ns/op
BenchmarkKVUpdate10_Native-4   	    1000	   1596340 ns/op
BenchmarkKVUpdate10_SQL-4      	    1000	   1668788 ns/op
BenchmarkKVUpdate100_Native-4  	     300	   5699316 ns/op
BenchmarkKVUpdate100_SQL-4     	     200	   7261698 ns/op
BenchmarkKVUpdate1000_Native-4 	      30	  45377174 ns/op
BenchmarkKVUpdate1000_SQL-4    	      30	  63683670 ns/op
BenchmarkKVUpdate10000_Native-4	       2	 505973716 ns/op
BenchmarkKVUpdate10000_SQL-4   	       2	 745406160 ns/op
BenchmarkKVDelete1_Native-4    	    2000	    853446 ns/op
BenchmarkKVDelete1_SQL-4       	    2000	    674473 ns/op
BenchmarkKVDelete10_Native-4   	    2000	   1126234 ns/op
BenchmarkKVDelete10_SQL-4      	    2000	   1045536 ns/op
BenchmarkKVDelete100_Native-4  	     500	   3686083 ns/op
BenchmarkKVDelete100_SQL-4     	     300	   4100471 ns/op
BenchmarkKVDelete1000_Native-4 	     100	  30876116 ns/op
BenchmarkKVDelete1000_SQL-4    	      50	  35879515 ns/op
BenchmarkKVDelete10000_Native-4	       5	 298790199 ns/op
BenchmarkKVDelete10000_SQL-4   	       2	 619428898 ns/op
BenchmarkKVScan1_Native-4      	    5000	    290198 ns/op
BenchmarkKVScan1_SQL-4         	    5000	    415768 ns/op
BenchmarkKVScan10_Native-4     	    5000	    327269 ns/op
BenchmarkKVScan10_SQL-4        	    3000	    511777 ns/op
BenchmarkKVScan100_Native-4    	    2000	    611106 ns/op
BenchmarkKVScan100_SQL-4       	    1000	   1329626 ns/op
BenchmarkKVScan1000_Native-4   	     500	   3263260 ns/op
BenchmarkKVScan1000_SQL-4      	     200	   8956545 ns/op
BenchmarkKVScan10000_Native-4  	      50	  33765755 ns/op
BenchmarkKVScan10000_SQL-4     	      20	  87925350 ns/op
```

<!-- Reviewable:start -->
---
This change is [<img src="https://reviewable.io/review_button.svg" height="35" align="absmiddle" alt="Reviewable"/>](https://reviewable.io/reviews/cockroachdb/cockroach/5931)
<!-- Reviewable:end -->