@petermattis 2016-04-07
<img class="emoji" title=":lgtm:" alt=":lgtm:" align="absmiddle" src="https://reviewable.io/lgtm.png" height="20" width="61"/> 

I bet the performance difference is due to the SQL sentinel key. Every row in the `bench.kv` table is 2 keys, while the kv-native stuff is only using a single key. Btw, you should almost always run the benchmarks 5-10 times (`go test -count 10`) and then format the data using `benchstat`.

---

Review status: 0 of 1 files reviewed at latest revision, 1 unresolved discussion.

---

*[sql/kv_test.go, line 284 \[r1\]](https://reviewable.io:443/reviews/cockroachdb/cockroach/5931#-KEm6PEdJ8JFvflcxZ9C:-KEm6PEdJ8JFvflcxZ9D:-292178108) ([raw file](https://github.com/cockroachdb/cockroach/blob/19be794ed2ea040a4cb78fc326426438d81fb468/sql/kv_test.go#L284)):*
You could pull this switch out of the loop:

```
var opFn func(count, run int) error
switch op {
case "insert":
  opFn = kv.insert
case "update":
  opFn = kv.update
...
}

for i := 0; i < b.N; i++ {
  if err := opFn(rows, i); err != nil {
    b.Fatal(err)
  }
}
```

---


*Comments from [Reviewable](https://reviewable.io:443/reviews/cockroachdb/cockroach/5931#-:-KEm5DaXWYVkg92SBYWs:1440641380)*
<!-- Sent from Reviewable.io -->