@JackKrupansky 2016-04-07
Cassandra has a feature called COMPACT STORAGE that effectively gives you raw KV without any additional overhead, so that one logical row translates into one cell in the storage layer.