@petermattis 2016-04-08
Here are GCE comparisons for the code in this PR (old is KV, new is SQL):

```
name              old time/op  new time/op   delta
KVInsert1-32      1.25ms ± 1%   0.93ms ± 1%   -25.72%   (p=0.000 n=10+9)
KVInsert10-32     1.61ms ± 1%   1.73ms ± 2%    +7.12%  (p=0.000 n=10+10)
KVInsert100-32    4.63ms ± 1%   9.31ms ± 2%  +101.05%  (p=0.000 n=10+10)
KVInsert1000-32   34.7ms ± 2%   93.0ms ± 4%  +167.92%  (p=0.000 n=10+10)
KVInsert10000-32   389ms ±10%   1356ms ±10%  +248.42%  (p=0.000 n=10+10)
KVUpdate1-32      1.65ms ± 1%   1.44ms ± 2%   -12.42%  (p=0.000 n=10+10)
KVUpdate10-32     2.39ms ± 1%   2.31ms ± 1%    -3.27%  (p=0.000 n=10+10)
KVUpdate100-32    7.69ms ± 1%   9.56ms ± 2%   +24.34%   (p=0.000 n=10+9)
KVUpdate1000-32   60.9ms ± 2%   83.2ms ± 4%   +36.73%   (p=0.000 n=9+10)
KVUpdate10000-32   661ms ± 6%   1001ms ± 8%   +51.41%  (p=0.000 n=10+10)
KVDelete1-32      1.23ms ± 1%   0.94ms ± 1%   -23.56%   (p=0.000 n=10+8)
KVDelete10-32     1.71ms ± 1%   1.44ms ± 2%   -16.08%  (p=0.000 n=10+10)
KVDelete100-32    5.13ms ± 2%   5.33ms ± 3%    +4.01%  (p=0.000 n=10+10)
KVDelete1000-32   39.3ms ± 3%   46.4ms ± 3%   +18.01%  (p=0.000 n=10+10)
KVDelete10000-32   404ms ± 6%   1057ms ± 4%  +161.48%  (p=0.000 n=10+10)
KVScan1-32         381µs ± 3%    534µs ± 5%   +40.16%  (p=0.000 n=10+10)
KVScan10-32        443µs ± 1%    720µs ± 3%   +62.61%   (p=0.000 n=8+10)
KVScan100-32       848µs ± 1%   1772µs ± 1%  +108.90%    (p=0.000 n=9+9)
KVScan1000-32     4.33ms ± 2%  11.57ms ± 3%  +166.94%  (p=0.000 n=10+10)
KVScan10000-32    44.0ms ± 4%  115.4ms ± 3%  +162.22%  (p=0.000 n=10+10)
```

For comparison, with the suggested change to use `PRIMARY KEY (k, v)` we get:

```
name              old time/op  new time/op  delta
KVInsert1-32      1.25ms ± 1%  0.89ms ± 2%  -29.17%  (p=0.000 n=10+10)
KVInsert10-32     1.61ms ± 1%  1.35ms ± 1%  -16.26%  (p=0.000 n=10+10)
KVInsert100-32    4.63ms ± 1%  5.30ms ± 2%  +14.40%  (p=0.000 n=10+10)
KVInsert1000-32   34.7ms ± 2%  46.1ms ± 3%  +32.85%  (p=0.000 n=10+10)
KVInsert10000-32   389ms ±10%   584ms ± 8%  +49.94%  (p=0.000 n=10+10)
KVDelete1-32      1.23ms ± 1%  0.96ms ± 2%  -22.65%   (p=0.000 n=10+8)
KVDelete10-32     1.71ms ± 1%  1.40ms ± 2%  -18.31%  (p=0.000 n=10+10)
KVDelete100-32    5.13ms ± 2%  4.87ms ± 3%   -5.13%  (p=0.000 n=10+10)
KVDelete1000-32   39.3ms ± 3%  39.8ms ± 2%     ~     (p=0.075 n=10+10)
KVDelete10000-32   404ms ± 6%   609ms ± 5%  +50.77%  (p=0.000 n=10+10)
KVScan1-32         381µs ± 3%   541µs ± 2%  +41.94%   (p=0.000 n=10+9)
KVScan10-32        443µs ± 1%   647µs ± 2%  +46.08%   (p=0.000 n=8+10)
KVScan100-32       848µs ± 1%  1317µs ± 3%  +55.25%   (p=0.000 n=9+10)
KVScan1000-32     4.33ms ± 2%  6.68ms ± 3%  +54.24%  (p=0.000 n=10+10)
KVScan10000-32    44.0ms ± 4%  64.4ms ± 3%  +46.47%  (p=0.000 n=10+10)
```

Curious that `Insert` starts out faster for KV-SQL, but scales poorly. Similar situation for `Delete`. Might be worthwhile investing some time to figure out what is going on there.