@spencerkimball 2016-04-08
```
name                    time/op
KVInsert1_Native-4       825µs ± 2%
KVInsert1_SQL-4          652µs ± 4%
KVInsert10_Native-4     1.18ms ± 7%
KVInsert10_SQL-4        1.37ms ± 5%
KVInsert100_Native-4    3.44ms ± 5%
KVInsert100_SQL-4       7.53ms ± 9%
KVInsert1000_Native-4   29.4ms ± 3%
KVInsert1000_SQL-4      82.8ms ± 4%
KVInsert10000_Native-4   374ms ± 8%
KVInsert10000_SQL-4      1.19s ± 7%
KVUpdate1_Native-4      1.22ms ± 1%
KVUpdate1_SQL-4         1.17ms ±10%
KVUpdate10_Native-4     1.60ms ± 1%
KVUpdate10_SQL-4        1.66ms ± 0%
KVUpdate100_Native-4    5.50ms ± 1%
KVUpdate100_SQL-4       6.97ms ± 2%
KVUpdate1000_Native-4   43.5ms ± 2%
KVUpdate1000_SQL-4      60.2ms ± 1%
KVUpdate10000_Native-4   491ms ± 3%
KVUpdate10000_SQL-4      744ms ± 4%
KVDelete1_Native-4       840µs ± 1%
KVDelete1_SQL-4          670µs ± 1%
KVDelete10_Native-4     1.14ms ± 1%
KVDelete10_SQL-4        1.15ms ±11%
KVDelete100_Native-4    4.21ms ± 3%
KVDelete100_SQL-4       4.54ms ± 1%
KVDelete1000_Native-4   31.8ms ± 5%
KVDelete1000_SQL-4      39.1ms ± 4%
KVDelete10000_Native-4   339ms ± 2%
KVDelete10000_SQL-4      655ms ± 5%
KVScan1_Native-4         345µs ± 4%
KVScan1_SQL-4            504µs ± 2%
KVScan10_Native-4        393µs ± 2%
KVScan10_SQL-4           608µs ± 3%
KVScan100_Native-4       702µs ± 4%
KVScan100_SQL-4         1.52ms ± 2%
KVScan1000_Native-4     3.64ms ± 7%
KVScan1000_SQL-4        10.4ms ± 4%
KVScan10000_Native-4    39.7ms ± 3%
KVScan10000_SQL-4        107ms ± 6%
```