@spencerkimball 2016-04-08
@petermattis the reason the KV was slower was that it wasn't making use of 1PC txns (it was letting the transaction auto commit). With 1PC, the native KV (without your key+value primary key change) is 30% faster than the SQL versions, as you'd expect.